Making programs executable, interactive, and implementing test()
The shebang line at the top of each file assumes that lua5.3 was installed from source according to the default instructions at lua.org
If running the demos using a different install location, then each file will need to be called manually.

In order to facilitate interactive exploration, I've adopted a pattern where each program defines functions, but has a test() function for the user to see or explore a simple demonstration.